Rewritten significant parts of the whole tool, all parsers wrok properly now.
The focus of this second iteration was primarily to fully extract all information form any mod.

The architecure is also set up in a way where is should be easy to incoperate into the game in the future, and integrate with any error reporting there.
The ny system tries to parse as much as possible, while reproting errors along the way.
After that a validation step can be called, to remove any invalid values and report issues with thos.

Issue reporting works via a callback, that gets fead errors as tehy happen.
Errors are represented by a tagged enum that can be processed as desired. That way it is also possible to easily provide localization for errors, as s liekly required by the game.
Instead of the tagged union we could also use Exception supertying (not throwing them, just passing them to the callback), I just wanted to try out the current approach.

Oh, also this now has some tests.
They use actual mods, but ive stripped aut the assets from the zips as those are not of interest to us and only bloat the repository size.

It is not required to copy all the dependencies that get emmited from including the cs parser, if found that only

Microsoft.CodeAnalysis.CSharp.dll
Microsoft.CodeAnalysis.dll
System.Collections.Immutable.dll
System.Runtime.CompilerServices.Unsafe.dll
System.Memory.dll

are required.

Once we move to dotnet8 we can also just build this as a single executable, avoiding the clutter.